requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Werkzeug==2.0.1
itsdangerous==2.0.1
python-dotenv==0.19.0
SQLAlchemy==1.4.23
oauth2client==4.1.3
google-auth==2.3.0
google-auth-oauthlib==0.4.6
gunicorn==21.2.0
flask-session==0.4.1
sendgrid==6.9.7
psycopg2-binary==2.9.9
greenlet==3.0.1
psycopg2==2.9.9
pyOpenSSL==23.2.0
auth_client==1.0.0
